updated logic for stock status
The logic for the stock status wasn’t handling edge cases all the way so I went down the route of making a reusable function (which might have been overkill), and then re-applying it across the ProductCard component.

It gave me a reason to write a unit test which I don’t have a ton of experience doing recently.

Along the way decided to have getStockColor follow the same essential logic for consistency.